+++ src/Sql/Sql/Properties/Resources.Designer.cs
@@ -574,6 +574,24 @@ namespace Microsoft.Azure.Commands.Sql.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Selected value for backup storage redundancy is geo-redundant storage. Note that database backups will be geo-replicated to the paired region. To learn more about Azure Paired Regions visit https://aka.ms/micreate-ragrs-regions..
+        /// </summary>
+        internal static string GeoBackupRedundancyChoosenWarning {
+            get {
+                return ResourceManager.GetString("GeoBackupRedundancyChoosenWarning", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to You have not specified the value for backup storage redundancy which will default to geo-redundant storage. Note that database backups will be geo-replicated to the paired region. To learn more about Azure Paired Regions visit https://aka.ms/micreate-ragrs-regions..
+        /// </summary>
+        internal static string GeoBackupRedundancyNotChoosenWarning {
+            get {
+                return ResourceManager.GetString("GeoBackupRedundancyNotChoosenWarning", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to The Active Directory Group &apos;{0}&apos; is not security enabled. Only Azure Active Directory Security Enabled Groups are supported..
         /// </summary>
         internal static string InvalidADGroupNotSecurity {