+++ src/Sql/Sql/ChangeLog.md
@@ -21,6 +21,7 @@
 * Added BackupStorageRedundancy to `New-AzSqlInstance` and `Get-AzSqlInstance`
 * Added cmdlet `Get-AzSqlServerActiveDirectoryOnlyAuthentication`
 * Added cmdlet `Enable-AzSqlServerActiveDirectoryOnlyAuthentication`
+* Added Force parameter to `New-AzSqlInstance`
 * Added cmdlets for Managed Database Log Replay service
 	- `Start-AzSqlInstanceDatabaseLogReplay`
 	- `Get-AzSqlInstanceDatabaseLogReplay`